requirements.txt
redis==4.5.5
tqdm==4.65.0
psutil==5.9.5
mysql-connector-python==8.0.33
mysql-replication==0.44
pymongo==4.3.3